requirements.txt
astronomer-cosmos
duckdb==1.4.4
pandas==2.2.3
apache-airflow-providers-postgres
astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy